requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.4
cron-descriptor==1.4.5
decorator==5.1.1
Django==4.2.16
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-multiselectfield==0.1.13
django-stubs==5.1.0
django-stubs-ext==5.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
eventlet==0.37.0
executing==2.1.0
filelock==3.16.1
flake8==7.1.1
future==1.0.0
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
ipython==8.29.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
matplotlib-inline==0.1.7
mccabe==0.7.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tldextract==5.1.3
traitlets==5.14.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.1.1
